requirements.txt
playwright==1.42.0
python-dotenv==1.0.1
pocketbase==0.9.2
requests==2.31.0
urllib3==2.2.1
python-dateutil==2.8.2
aiohttp 